Check if $image->pid is set

+++ inc/3rd-party/nextgen-gallery/inc/classes/class-ngg-storage.php
@@ -2,7 +2,14 @@
 defined( 'ABSPATH' ) or die( 'Cheatin\' uh?' );
 
 class Imagify_NGG_Storage extends Mixin {
-    
+    /**
+     * Recover image from backup copy and reprocess it
+     *
+	 * @since 1.5
+	 * @author Jonathan Buttigieg
+	 *
+     * @return void
+     **/
     function recover_image( $image ) {
         // $image could be an object or an (int) image ID
         if ( is_numeric( $image ) ) {
@@ -10,7 +17,9 @@ class Imagify_NGG_Storage extends Mixin {
         }
         
 		// Remove Imagify data
-		Imagify_NGG_DB()->delete( $image->pid );
+		if ( isset( $image->pid ) ) {
+			Imagify_NGG_DB()->delete( $image->pid );	
+		}
 		
 		return $this->call_parent( 'recover_image', $image );
     }
